server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
motor==3.5.1
pymongo==4.5
python-dotenv==1.0.1
pyjwt==2.9.0
pydantic==2.9.2
websockets==13.0.1
passlib==1.7.4